Add paragraph on redirects
When a spec transitions to the W3C organization, it usually means it is somewhat
"successful", in other words, that people are talking about it, which they
usually do online by linking to the spec.

The repo transition kills the `https://[org].github.io/[repo]` link by default.
Each time a link breaks, some web pioneer suffers. This update explains how to
put redirects in place.